requirements.txt
requests
Pillow
beautifulsoup4
transformers
gradio